Trying to make sense by moving conodes around...
