+++ app/conode/client.go
@@ -17,12 +17,13 @@ import (
 	"github.com/dedis/cothority/lib/hashid"
 	"github.com/dedis/cothority/lib/logutils"
 	"github.com/dedis/cothority/lib/app"
+	"github.com/dedis/cothority/lib/conode"
 )
 
 var muStats sync.Mutex
 
 var MAX_N_SECONDS int = 1 * 60 * 60 // 1 hours' worth of seconds
-var MAX_N_ROUNDS int = MAX_N_SECONDS / int(ROUND_TIME / time.Second)
+var MAX_N_ROUNDS int = MAX_N_SECONDS / int(conode.ROUND_TIME / time.Second)
 
 func RunClient(flags *app.Flags, conf *app.ConfigColl){
 	dbg.Lvl4("Starting to run stampclient")
