--- app/conode/constants.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"time"
-
-	"github.com/dedis/cothority/proto/sign"
-)
-
-// time we wait between rounds
-var ROUND_TIME time.Duration = sign.ROUND_TIME
